+++ packages/core/src/bin/utils/startIsolated.ts
@@ -8,6 +8,7 @@ import { promiseWithResolvers } from "@/utils/promiseWithResolvers.js";
 import { isTable, sql } from "drizzle-orm";
 import type { PonderApp } from "../commands/start.js";
 import type { CliOptions } from "../ponder.js";
+import { runCodegen } from "./codegen.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -37,6 +38,9 @@ export async function startIsolated(
   }: PonderApp,
   cliOptions: CliOptions,
 ) {
+  await database.migrateSync();
+  runCodegen({ common });
+
   const appState: { [chainName: string]: WorkerInfo } = {};
   const workerPath = join(__dirname, "..", "..", "runtime", "isolated.js");