+++ packages/core/src/database/index.test.ts
@@ -1,4 +1,5 @@
 import { setupCommon, setupIsolatedDatabase } from "@/_test/setup.js";
+import { getChain } from "@/_test/utils.js";
 import { buildSchema } from "@/build/schema.js";
 import { onchainEnum, onchainTable, primaryKey } from "@/drizzle/onchain.js";
 import { createRealtimeIndexingStore } from "@/indexing-store/realtime.js";
@@ -66,7 +67,12 @@ test("migrate() succeeds with empty schema", async (context) => {
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const tableNames = await getUserTableNames(database, "public");
   expect(tableNames).toContain("account");
@@ -120,7 +126,12 @@ test("migrate() with empty schema creates tables and enums", async (context) => 
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const tableNames = await getUserTableNames(database, "public");
   expect(tableNames).toContain("account");
@@ -150,7 +161,12 @@ test("migrate() throws with schema used", async (context) => {
         .statements,
     },
   });
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
   await context.common.shutdown.kill();
 
   context.common.shutdown = createShutdown();
@@ -172,7 +188,12 @@ test("migrate() throws with schema used", async (context) => {
   });
 
   const error = await databaseTwo
-    .migrate({ buildId: "def", ordering: "multichain" })
+    .migrate({
+      buildId: "def",
+      chains: [],
+      finalizedBlocks: [],
+      ordering: "multichain",
+    })
     .catch((err) => err);
 
   expect(error).toBeDefined();
@@ -203,7 +224,12 @@ test("migrate() throws with schema used after waiting for lock", async (context)
         .statements,
     },
   });
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const databaseTwo = await createDatabase({
     common: context.common,
@@ -222,7 +248,12 @@ test("migrate() throws with schema used after waiting for lock", async (context)
   });
 
   const error = await databaseTwo
-    .migrate({ buildId: "abc", ordering: "multichain" })
+    .migrate({
+      buildId: "abc",
+      chains: [],
+      finalizedBlocks: [],
+      ordering: "multichain",
+    })
     .catch((err) => err);
 
   expect(error).toBeDefined();
@@ -247,7 +278,12 @@ test("migrate() succeeds with crash recovery", async (context) => {
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
   await context.common.shutdown.kill();
 
   context.common.shutdown = createShutdown();
@@ -268,7 +304,12 @@ test("migrate() succeeds with crash recovery", async (context) => {
     },
   });
 
-  await databaseTwo.migrate({ buildId: "abc", ordering: "multichain" });
+  await databaseTwo.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const metadata = await databaseTwo.userQB.wrap((db) =>
     db.select().from(getPonderMetaTable("public")),
@@ -303,7 +344,12 @@ test("migrate() succeeds with crash recovery after waiting for lock", async (con
         .statements,
     },
   });
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const databaseTwo = await createDatabase({
     common: context.common,
@@ -321,7 +367,12 @@ test("migrate() succeeds with crash recovery after waiting for lock", async (con
     },
   });
 
-  await databaseTwo.migrate({ buildId: "abc", ordering: "multichain" });
+  await databaseTwo.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   await context.common.shutdown.kill();
 });
@@ -403,11 +454,14 @@ test("migrate() with crash recovery reverts rows", async (context) => {
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   // setup tables, reorg tables, and metadata checkpoint
-
-  await createTriggers(database.userQB, { tables: [account] });
 
   await database.userQB.wrap((db) =>
     db.update(getPonderMetaTable("public")).set({
@@ -429,8 +483,10 @@ test("migrate() with crash recovery reverts rows", async (context) => {
   await commitBlock(database.userQB, {
     checkpoint: createCheckpoint({ chainId: 1n, blockNumber: 9n }),
     table: account,
-    ordering: "multichain",
+    preBuild: { ordering: "multichain" },
   });
+
+  await createTriggers(database.userQB, { tables: [account] });
 
   await indexingStore
     .insert(account)
@@ -439,7 +495,7 @@ test("migrate() with crash recovery reverts rows", async (context) => {
   await commitBlock(database.userQB, {
     checkpoint: createCheckpoint({ chainId: 1n, blockNumber: 11n }),
     table: account,
-    ordering: "multichain",
+    preBuild: { ordering: "multichain" },
   });
 
   await database.userQB.wrap((db) =>
@@ -447,6 +503,7 @@ test("migrate() with crash recovery reverts rows", async (context) => {
       chainId: 1,
       chainName: "mainnet",
       latestCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
+      finalizedCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
       safeCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
     }),
   );
@@ -473,6 +530,15 @@ test("migrate() with crash recovery reverts rows", async (context) => {
 
   const checkpoint = await databaseTwo.migrate({
     buildId: "abc",
+    chains: [getChain()],
+    finalizedBlocks: [
+      {
+        timestamp: "0x1",
+        number: "0xa",
+        hash: "0x",
+        parentHash: "0x",
+      },
+    ],
     ordering: "multichain",
   });
 
@@ -529,7 +595,12 @@ test("migrate() with crash recovery drops indexes and triggers", async (context)
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   await createIndexes(database.userQB, {
     statements: buildSchema({ schema: { account }, ordering: "multichain" })
@@ -541,6 +612,7 @@ test("migrate() with crash recovery drops indexes and triggers", async (context)
       chainId: 1,
       chainName: "mainnet",
       latestCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
+      finalizedCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
       safeCheckpoint: createCheckpoint({ chainId: 1n, blockNumber: 10n }),
     }),
   );
@@ -571,7 +643,19 @@ test("migrate() with crash recovery drops indexes and triggers", async (context)
     },
   });
 
-  await databaseTwo.migrate({ buildId: "abc", ordering: "multichain" });
+  await databaseTwo.migrate({
+    buildId: "abc",
+    chains: [getChain()],
+    finalizedBlocks: [
+      {
+        timestamp: "0x1",
+        number: "0xa",
+        hash: "0x",
+        parentHash: "0x",
+      },
+    ],
+    ordering: "multichain",
+  });
 
   const indexNames = await getUserIndexNames(databaseTwo, "public", "account");
 
@@ -600,7 +684,12 @@ test("heartbeat updates the heartbeat_at value", async (context) => {
     },
   });
 
-  await database.migrate({ buildId: "abc", ordering: "multichain" });
+  await database.migrate({
+    buildId: "abc",
+    chains: [],
+    finalizedBlocks: [],
+    ordering: "multichain",
+  });
 
   const row = await database.userQB.wrap((db) =>
     db