+++ packages/core/src/runtime/isolated.ts
@@ -334,12 +334,14 @@ export async function runIsolated({
             chainName: chain.name,
             chainId: chain.id,
             latestCheckpoint: syncProgress.getCheckpoint({ tag: "start" }),
+            finalizedCheckpoint: syncProgress.getCheckpoint({ tag: "start" }),
             safeCheckpoint: syncProgress.getCheckpoint({ tag: "start" }),
           })
           .onConflictDoUpdate({
             target: PONDER_CHECKPOINT.chainName,
             set: {
               safeCheckpoint: sql`excluded.safe_checkpoint`,
+              finalizedCheckpoint: sql`excluded.finalized_checkpoint`,
               latestCheckpoint: sql`excluded.latest_checkpoint`,
             },
           }),
@@ -471,12 +473,14 @@ export async function runIsolated({
                 chainName: chain.name,
                 chainId: chain.id,
                 latestCheckpoint: events.checkpoint,
+                finalizedCheckpoint: events.checkpoint,
                 safeCheckpoint: events.checkpoint,
               })
               .onConflictDoUpdate({
                 target: PONDER_CHECKPOINT.chainName,
                 set: {
                   safeCheckpoint: sql`excluded.safe_checkpoint`,
+                  finalizedCheckpoint: sql`excluded.finalized_checkpoint`,
                   latestCheckpoint: sql`excluded.latest_checkpoint`,
                 },
               }),
@@ -590,7 +594,7 @@ export async function runIsolated({
                     commitBlock(tx, {
                       table,
                       checkpoint,
-                      ordering: preBuild.ordering,
+                      preBuild,
                     }),
                   ),
                 );
@@ -634,7 +638,7 @@ export async function runIsolated({
           const counts = await revert(tx, {
             tables,
             checkpoint: event.checkpoint,
-            ordering: preBuild.ordering,
+            preBuild,
           });
 
           for (const [index, table] of tables.entries()) {
@@ -648,39 +652,21 @@ export async function runIsolated({
         });
 
         break;
-      case "finalize":
-        await database.userQB.transaction(async (tx) => {
-          await tx.wrap((tx) =>
-            tx
-              .update(PONDER_CHECKPOINT)
-              .set({
-                safeCheckpoint: event.checkpoint,
-              })
-              .where(eq(PONDER_CHECKPOINT.chainName, event.chain.name)),
-          );
+      case "finalize": {
+        const count = await finalize(database.userQB, {
+          checkpoint: event.checkpoint,
+          tables,
+          preBuild,
+          namespaceBuild,
+        });
 
-          const counts = await finalize(tx, {
-            tables,
-            checkpoint: event.checkpoint,
-            ordering: preBuild.ordering,
-          });
-
-          for (const [index, table] of tables.entries()) {
-            common.logger.info({
-              service: "database",
-              msg: `Finalized ${counts[index]} operations from '${getTableName(table)}'`,
-            });
-          }
-
-          const decoded = decodeCheckpoint(event.checkpoint);
-
-          common.logger.debug({
-            service: "database",
-            msg: `Updated finalized checkpoint to (timestamp=${decoded.blockTimestamp} chainId=${decoded.chainId} block=${decoded.blockNumber})`,
-          });
+        common.logger.info({
+          service: "database",
+          msg: `Finalized ${count} operations.`,
         });
 
         break;
+      }
       default:
         never(event);
     }